requirements.txt
pdfminer.six
numpy